Created bullet_class
created a bullet class that creates a projectile that moves across the screen in the direction of where the mouse was clicked. checks for collisions between itself and the walkers. removes both if it collides. Stops the player from moving once fired and switches to the enemies turn once it is destroyed. It is also destroyed if it goes off the screen.